Update simplification project for double

+++ jsplat-simplification/src/main/java/de/javagl/jsplat/simplification/incremental/Edge.java
@@ -46,7 +46,7 @@ public class Edge
     /**
      * The weight
      */
-    public float weight;
+    public double weight;
 
     /**
      * Creates a new instance
