+++ jsplat-simplification/src/main/java/de/javagl/jsplat/simplification/incremental/nanogs/NanoGsMath.java
@@ -24,14 +24,14 @@ class NanoGsMath
     /**
      * A thread-local 3x3 matrix for "R"
      */
-    private static final ThreadLocal<float[]> THREAD_LOCAL_R =
-        ThreadLocal.withInitial(() -> new float[9]);        
+    private static final ThreadLocal<double[]> THREAD_LOCAL_R =
+        ThreadLocal.withInitial(() -> new double[9]);        
 
     /**
      * A thread-local 3x3 matrix for "matrix"
      */
-    private static final ThreadLocal<float[]> THREAD_LOCAL_matrix =
-        ThreadLocal.withInitial(() -> new float[9]);        
+    private static final ThreadLocal<double[]> THREAD_LOCAL_matrix =
+        ThreadLocal.withInitial(() -> new double[9]);        
 
     /**
      * A container for the results of an eigendecomposition
@@ -41,12 +41,12 @@ class NanoGsMath
         /**
          * The eigenvalues
          */
-        final float[] eigenvalues = new float[3];
+        final double[] eigenvalues = new double[3];
 
         /**
          * The eigenvectors, as a flat 9-element array
          */
-        final float[] eigenvectors = new float[9];
+        final double[] eigenvectors = new double[9];
     }
 
     /**
@@ -55,17 +55,17 @@ class NanoGsMath
      * @param A The input array
      * @return The determinant
      */
-    static float det3Flat(float A[])
+    static double det3Flat(double A[])
     {
-        float a00 = A[0];
-        float a01 = A[1];
-        float a02 = A[2];
-        float a10 = A[3];
-        float a11 = A[4];
-        float a12 = A[5];
-        float a20 = A[6];
-        float a21 = A[7];
-        float a22 = A[8];
+        double a00 = A[0];
+        double a01 = A[1];
+        double a02 = A[2];
+        double a10 = A[3];
+        double a11 = A[4];
+        double a12 = A[5];
+        double a20 = A[6];
+        double a21 = A[7];
+        double a22 = A[8];
         return (a00 * (a11 * a22 - a12 * a21) - a01 * (a10 * a22 - a12 * a20)
             + a02 * (a10 * a21 - a11 * a20));
     }
@@ -77,45 +77,45 @@ class NanoGsMath
      * @param R The array
      * @param q The output
      */
-    static void rotmatToQuatFlat(float R[], float q[])
+    static void rotmatToQuatFlat(double R[], double q[])
     {
-        float m00 = R[0];
-        float m11 = R[4];
-        float m22 = R[8];
-        float tr = m00 + m11 + m22;
-        float qw, qx, qy, qz;
+        double m00 = R[0];
+        double m11 = R[4];
+        double m22 = R[8];
+        double tr = m00 + m11 + m22;
+        double qw, qx, qy, qz;
 
         if (tr > 0)
         {
-            float S = (float) Math.sqrt(tr + 1.0) * 2.0f;
-            qw = 0.25f * S;
+            double S = Math.sqrt(tr + 1.0) * 2.0;
+            qw = 0.25 * S;
             qx = (R[7] - R[5]) / S;
             qy = (R[2] - R[6]) / S;
             qz = (R[3] - R[1]) / S;
         }
         else if (m00 > m11 && m00 > m22)
         {
-            float S = (float) Math.sqrt(1.0 + m00 - m11 - m22) * 2.0f;
+            double S = Math.sqrt(1.0 + m00 - m11 - m22) * 2.0;
             qw = (R[7] - R[5]) / S;
-            qx = 0.25f * S;
+            qx = 0.25 * S;
             qy = (R[1] + R[3]) / S;
             qz = (R[2] + R[6]) / S;
         }
         else if (m11 > m22)
         {
-            float S = (float) Math.sqrt(1.0 + m11 - m00 - m22) * 2.0f;
+            double S = Math.sqrt(1.0 + m11 - m00 - m22) * 2.0;
             qw = (R[2] - R[6]) / S;
             qx = (R[1] + R[3]) / S;
-            qy = 0.25f * S;
+            qy = 0.25 * S;
             qz = (R[5] + R[7]) / S;
         }
         else
         {
-            float S = (float) Math.sqrt(1.0 + m22 - m00 - m11) * 2.0f;
+            double S = Math.sqrt(1.0 + m22 - m00 - m11) * 2.0;
             qw = (R[3] - R[1]) / S;
             qx = (R[2] + R[6]) / S;
             qy = (R[5] + R[7]) / S;
-            qz = 0.25f * S;
+            qz = 0.25 * S;
         }
         q[0] = qw;
         q[1] = qx;
@@ -129,12 +129,12 @@ class NanoGsMath
      * 
      * @param q The quaternion
      */
-    private static void normalize(float[] q)
+    private static void normalize(double[] q)
     {
-        float lenSq = q[0] * q[0] + q[1] * q[1] + q[2] * q[2] + q[3] * q[3];
+        double lenSq = q[0] * q[0] + q[1] * q[1] + q[2] * q[2] + q[3] * q[3];
         if (lenSq > 0)
         {
-            float len = (float) Math.sqrt(lenSq);
+            double len = Math.sqrt(lenSq);
             for (int i = 0; i < 4; i++)
             {
                 q[i] /= len;
@@ -152,7 +152,7 @@ class NanoGsMath
      * @param dst The target array
      * @param dOffset The target offset
      */
-    static void transpose3Into(float src[], int sOffset, float dst[],
+    static void transpose3Into(double src[], int sOffset, double dst[],
         int dOffset)
     {
         dst[dOffset + 0] = src[sOffset + 0];
@@ -179,10 +179,10 @@ class NanoGsMath
      * @param sz The scale value along z
      * @param out The output array
      */
-    static void sigmaFromQuatScaleFlatInto(float qw, float qx, float qy,
-        float qz, float sx, float sy, float sz, float[] out)
+    static void sigmaFromQuatScaleFlatInto(double qw, double qx, double qy,
+        double qz, double sx, double sy, double sz, double[] out)
     {
-        float[] R = THREAD_LOCAL_R.get();
+        double[] R = THREAD_LOCAL_R.get();
         quatToRotmatInto(qw, qx, qy, qz, R, 0);
         sigmaFromRotVarInto(R, 0, sx * sx, sy * sy, sz * sz, out, 0);
     }
@@ -198,18 +198,18 @@ class NanoGsMath
      * @param out The output array
      * @param offset The offset
      */
-    static void quatToRotmatInto(float w, float x, float y, float z,
-        float out[], int offset)
+    static void quatToRotmatInto(double w, double x, double y, double z,
+        double out[], int offset)
     {
-        float xx = x * x;
-        float yy = y * y;
-        float zz = z * z;
-        float wx = w * x;
-        float wy = w * y;
-        float wz = w * z;
-        float xy = x * y;
-        float xz = x * z;
-        float yz = y * z;
+        double xx = x * x;
+        double yy = y * y;
+        double zz = z * z;
+        double wx = w * x;
+        double wy = w * y;
+        double wz = w * z;
+        double xy = x * y;
+        double xz = x * z;
+        double yz = y * z;
 
         out[offset + 0] = 1 - 2 * (yy + zz);
         out[offset + 1] = 2 * (xy - wz);
@@ -236,18 +236,18 @@ class NanoGsMath
      * @param out The output array
      * @param oOffset The offset where the output should be written
      */
-    static void sigmaFromRotVarInto(float R[], int rOffset, float vx, float vy,
-        float vz, float out[], int oOffset)
+    static void sigmaFromRotVarInto(double R[], int rOffset, double vx, double vy,
+        double vz, double out[], int oOffset)
     {
-        float r00 = R[rOffset + 0];
-        float r01 = R[rOffset + 1];
-        float r02 = R[rOffset + 2];
-        float r10 = R[rOffset + 3];
-        float r11 = R[rOffset + 4];
-        float r12 = R[rOffset + 5];
-        float r20 = R[rOffset + 6];
-        float r21 = R[rOffset + 7];
-        float r22 = R[rOffset + 8];
+        double r00 = R[rOffset + 0];
+        double r01 = R[rOffset + 1];
+        double r02 = R[rOffset + 2];
+        double r10 = R[rOffset + 3];
+        double r11 = R[rOffset + 4];
+        double r12 = R[rOffset + 5];
+        double r20 = R[rOffset + 6];
+        double r21 = R[rOffset + 7];
+        double r22 = R[rOffset + 8];
 
         out[oOffset + 0] = r00 * r00 * vx + r01 * r01 * vy + r02 * r02 * vz;
         out[oOffset + 1] = r00 * r10 * vx + r01 * r11 * vy + r02 * r12 * vz;
@@ -268,25 +268,25 @@ class NanoGsMath
      * @param matrix The matrix
      * @param result The result
      */
-    static void eigenSymmetric3x3Flat(float[] matrix, Eigendecomposition3x3 result)
+    static void eigenSymmetric3x3Flat(double[] matrix, Eigendecomposition3x3 result)
     {
-        float[] mat = THREAD_LOCAL_matrix.get();
+        double[] mat = THREAD_LOCAL_matrix.get();
         System.arraycopy(matrix, 0, mat, 0, 9);
-        float[] V = result.eigenvectors;
-        V[0] = 1.0f;
-        V[1] = 0.0f;
-        V[2] = 0.0f;
-        V[3] = 0.0f;
-        V[4] = 1.0f;
-        V[5] = 0.0f;
-        V[6] = 0.0f;
-        V[7] = 0.0f;
-        V[8] = 1.0f;
+        double[] V = result.eigenvectors;
+        V[0] = 1.0;
+        V[1] = 0.0;
+        V[2] = 0.0;
+        V[3] = 0.0;
+        V[4] = 1.0;
+        V[5] = 0.0;
+        V[6] = 0.0;
+        V[7] = 0.0;
+        V[8] = 1.0;
         for (int i = 0; i < 24; i++)
         {
             int p = 0;
             int q = 1;
-            float maxAbs = Math.abs(mat[1]);
+            double maxAbs = Math.abs(mat[1]);
 
             if (Math.abs(mat[2]) > maxAbs)
             {
@@ -311,15 +311,15 @@ class NanoGsMath
             int pq = 3 * p + q;
             int qp = 3 * q + p;
 
-            float app = mat[pp];
-            float aqq = mat[qq];
-            float apq = mat[pq];
+            double app = mat[pp];
+            double aqq = mat[qq];
+            double apq = mat[pq];
 
-            float tau = (aqq - app) / (2 * apq);
-            float t = (float) (Math.signum(tau)
+            double tau = (aqq - app) / (2 * apq);
+            double t = (Math.signum(tau)
                 / (Math.abs(tau) + Math.sqrt(1 + tau * tau)));
-            float c = 1 / (float) Math.sqrt(1 + t * t);
-            float s = t * c;
+            double c = 1 / Math.sqrt(1 + t * t);
+            double s = t * c;
 
             for (int k = 0; k < 3; k++)
             {
@@ -331,8 +331,8 @@ class NanoGsMath
                 int kq = 3 * k + q;
                 int pk = 3 * p + k;
                 int qk = 3 * q + k;
-                float akp = mat[kp];
-                float akq = mat[kq];
+                double akp = mat[kp];
+                double akq = mat[kq];
                 mat[kp] = c * akp - s * akq;
                 mat[pk] = mat[kp];
                 mat[kq] = s * akp + c * akq;
@@ -348,8 +348,8 @@ class NanoGsMath
             {
                 int kp = 3 * k + p;
                 int kq = 3 * k + q;
-                float vkp = V[kp];
-                float vkq = V[kq];
+                double vkp = V[kp];
+                double vkq = V[kq];
                 V[kp] = c * vkp - s * vkq;
                 V[kq] = s * vkp + c * vkq;
             }
@@ -367,17 +367,17 @@ class NanoGsMath
      * @param p The percentile
      * @return The result
      */
-    static float percentileInPlace(float xs[], float p)
+    static double percentileInPlace(double xs[], double p)
     {
         Arrays.parallelSort(xs);
         if (xs.length == 0)
         {
             return 0;
         }
-        float t = (xs.length - 1) * p;
+        double t = (xs.length - 1) * p;
         int i = (int) Math.floor(t);
         int j = Math.min(i + 1, xs.length - 1);
-        float w = t - i;
+        double w = t - i;
         return xs[i] * (1 - w) + xs[j] * w;
     }
